# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateQuizResponse:
    properties:
      quiz_id:
        type: string
    type: object
...
  dto.ReadQuizQuestionResponse:
    properties:
      question:
        type: string
      quiz_options:
        items:
          type: string
        type: array
      quiz_question_id:
        type: string
    type: object
  dto.ReadQuizResponse:
    properties:
      quiz_id:
        type: string
      quiz_questions:
        items:
          $ref: '#/definitions/dto.ReadQuizQuestionResponse'
        type: array
      quiz_title:
        type: string
    type: object
...
  dto.SolveQuestionRequest:
    properties:
      answer:
        items:
          type: string
        type: array
      quiz_question_id:
        type: string
    type: object
  dto.SolveQuestionResponse:
    properties:
      correct_fields:
        items:
          type: string
        type: array
      is_correct:
        type: boolean
      quiz_question_id:
        type: string
    type: object
  dto.SolveQuizRequest:
    properties:
      attempts:
        items:
          $ref: '#/definitions/dto.SolveQuestionRequest'
        type: array
    type: object
  dto.SolveQuizResponse:
    properties:
      is_correct:
        type: boolean
      questions_answers:
        items:
          $ref: '#/definitions/dto.SolveQuestionResponse'
        type: array
    type: object
...
  entity.Question:
    properties:
      answers:
        items:
          type: string
        type: array
      id:
        type: string
      options:
        items:
          type: string
        type: array
      question:
        type: string
      type:
        $ref: '#/definitions/model.QuizType'
    type: object
  entity.Quiz:
    properties:
      id:
        type: string
      questions:
        items:
          $ref: '#/definitions/entity.Question'
        type: array
      quiz_name:
        type: string
      team_id:
        type: string
      user_id:
        type: string
      user_team_id:
        type: string
    type: object
...
  model.QuizType:
    enum:
    - multiple_choice
    - true_false
    type: string
    x-enum-varnames:
    - MultipleChoice
    - TrueFalse
...
info:
  contact: {}
  title: StudyWithMe API
  version: "1.0"
paths:
...
  /quizzes:
    post:
      consumes:
      - application/json
      parameters:
      - description: The create quiz request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.Quiz'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateQuizResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create a new quiz
      tags:
      - quizzes
  /quizzes/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: The id for quiz
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Quiz'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get a quiz with answers
      tags:
      - quizzes
  /quizzes/{id}/test:
    get:
      consumes:
      - application/json
      parameters:
      - description: The id for quiz
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReadQuizResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get a quiz without answers
      tags:
      - quizzes
    post:
      consumes:
      - application/json
      parameters:
      - description: The solve quiz request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SolveQuizRequest'
      - description: The id for quiz
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SolveQuizResponse'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Solve a quiz
      tags:
      - quizzes
  /quizzes/team/{teamId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Team ID
        in: path
        name: teamId
        required: true
        type: string
      - description: Page size (default 10)
        in: query
        name: pageSize
        type: integer
      - description: Last key for pagination
        in: query
        name: lastKey
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get quizzes by team with pagination
      tags:
      - quizzes
  /quizzes/user/{userId}/team/{teamId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Team ID
        in: path
        name: teamId
        required: true
        type: string
      - description: Page size (default 10)
        in: query
        name: pageSize
        type: integer
      - description: Last key for pagination
        in: query
        name: lastKey
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get quizzes by user with pagination
      tags:
      - quizzes
...
securityDefinitions:
  Bearer:
    description: 'Bearer JWT token. Example: "Bearer YOUR_TOKEN"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"